fix note's controllers validations

+++ backend/controllers/note.controller.js
@@ -5,18 +5,37 @@ const pool = require('../config/db')
 const fetchNote = asyncHandler(async (req, res) => {
     const noteIdValue = await joiValidation.idSchema.validateAsync(req.params)
     const folderIdValue = await joiValidation.idSchema.validateAsync(req.body)
-    const {folderId} = folderIdValue
-    const {noteId} = noteIdValue
+    const folderId = folderIdValue.id
+    const noteId = noteIdValue.id
+    const {userId} = req.user
 
-    const [rows] = await pool.query(`
+    const [folder] = await pool.query(`
+        SELECT *
+        FROM folders
+        WHERE id=? AND user_id=?
+        `, [folderId, userId])
+    
+    if (folder.length === 0) {
+        const err = new Error('Folder not found')
+        err.status = 404
+        throw err
+    }
+
+    const [notes] = await pool.query(`
         SELECT * FROM notes
         WHERE folder_id=? AND id=?
         `, [folderId, noteId])
+    
+    if (notes.length === 0) {
+        const err = new Error('Note not found')
+        err.status = 404
+        throw err
+    }
 
     res.json({
         message: 'Fetch note',
         success: true,
-        data: rows
+        data: notes
     })
 })
 
@@ -41,6 +60,20 @@ const createNotes = asyncHandler(async (req, res) => {
 const deleteNote = asyncHandler(async (req, res) => {
     const value = await joiValidation.idSchema.validateAsync(req.params)
     const {id} = value
+    const {userId} = req.user
+
+    const [rows] = await pool.query(`
+        SELECT *
+        FROM notes
+        JOIN folders ON notes.folder_id = folders.id
+        WHERE notes.id=? and folders.user_id=?
+        `, [id, userId])
+
+    if (rows.length === 0) {
+        const err = new Error('Note not found')
+        err.status = 404
+        throw err
+    }
 
     await pool.query(`
         DELETE FROM notes
@@ -58,6 +91,20 @@ const updateNote = asyncHandler(async (req, res) => {
     const bodyValue = await joiValidation.noteSchema.validateAsync(req.body)
     const {id} = idValue
     const {title, content} = bodyValue
+    const {userId} = req.user
+
+    const [notes] = await pool.query(`
+        SELECT *
+        FROM notes
+        JOIN folders ON notes.folder_id = folders.id
+        WHERE notes.id=? AND folders.user_id=?
+        `, [id, userId])
+
+    if (notes.length === 0) {
+        const err = new Error('Note not found')
+        err.status = 404
+        throw err
+    }
 
     await pool.query(`
         UPDATE notes
